legacy/backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-filter==23.5

# Autenticación
djangorestframework-simplejwt==5.3.1
dj-rest-auth==5.0.2
django-allauth==0.60.0

# Base de Datos
psycopg2-binary==2.9.9

# Storage
Pillow==10.2.0
django-storages==1.14.2
boto3==1.34.0

# Utilidades
python-decouple==3.8
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==21.0.0

# Debug
django-debug-toolbar==4.2.0
